cortex/requirements.txt
# Telos Cortex - Python Dependencies

# gRPC Runtime
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Protocol Buffers
protobuf>=4.25.0

# Configuration
PyYAML>=6.0.1

# Utilities (optional but recommended)
